requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp-retry==2.9.1
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
exceptiongroup==1.3.0
Flask==3.1.1
frozenlist==1.7.0
h11==0.16.0
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.6.3
outcome==1.3.0.post0
pillow==11.3.0
propcache==0.3.2
PyJWT==2.10.1
PySocks==1.7.1
python-dotenv==1.1.1
requests==2.32.4
schedule==1.2.2
selenium==4.34.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.30.0
trio-websocket==0.12.2
twilio==9.6.3
typing_extensions==4.14.0
urllib3==2.4.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.20.1
zipp==3.23.0
